+++ include/ygm/container/bag.hpp
@@ -6,11 +6,11 @@
 #pragma once
 
 #include <cereal/archives/json.hpp>
+#include <initializer_list>
 #include <ygm/container/container_traits.hpp>
 #include <ygm/container/detail/base_async_insert.hpp>
-#include <ygm/container/detail/base_iteration.hpp>
-#include <ygm/container/detail/base_iteration_tools.hpp>
 #include <ygm/container/detail/base_count.hpp>
+#include <ygm/container/detail/base_iteration.hpp>
 #include <ygm/container/detail/base_misc.hpp>
 #include <ygm/container/detail/round_robin_partitioner.hpp>
 #include <ygm/random.hpp>
@@ -21,8 +21,7 @@ template <typename Item>
 class bag : public detail::base_async_insert_value<bag<Item>, std::tuple<Item>>,
             public detail::base_count<bag<Item>, std::tuple<Item>>,
             public detail::base_misc<bag<Item>, std::tuple<Item>>,
-            public detail::base_iteration<bag<Item>, std::tuple<Item>>,
-            public detail::base_iteration_tools<bag<Item>, std::tuple<Item>> {
+            public detail::base_iteration<bag<Item>, std::tuple<Item>> {
   friend class detail::base_misc<bag<Item>, std::tuple<Item>>;
 
  public:
@@ -35,9 +34,20 @@ class bag : public detail::base_async_insert_value<bag<Item>, std::tuple<Item>>,
   bag(ygm::comm &comm) : m_comm(comm), pthis(this), partitioner(comm) {
     pthis.check(m_comm);
   }
+
+  bag(ygm::comm &comm, std::initializer_list<Item> l)
+      : m_comm(comm), pthis(this), partitioner(comm) {
+    pthis.check(m_comm);
+    if (m_comm.rank0()) {
+      for (const Item &i : l) {
+        async_insert(i);
+      }
+    }
+  }
+
   ~bag() { m_comm.barrier(); }
 
-  bag(const self_type &other) // If I remove const it compiles
+  bag(const self_type &other)  // If I remove const it compiles
       : m_comm(other.comm()), pthis(this), partitioner(other.comm()) {
     pthis.check(m_comm);
   }
@@ -83,7 +93,7 @@ class bag : public detail::base_async_insert_value<bag<Item>, std::tuple<Item>>,
 
   size_t local_size() const { return m_local_bag.size(); }
 
-  size_t local_count(const value_type& val) const { 
+  size_t local_count(const value_type &val) const {
     return std::count(m_local_bag.begin(), m_local_bag.end(), val);
   }
