+++ include/ygm/container/detail/base_count.hpp
@@ -7,13 +7,13 @@
 
 #include <tuple>
 #include <utility>
+#include <ygm/collective.hpp>
 
 namespace ygm::container::detail {
 
 template <typename derived_type, typename for_all_args>
 struct base_count {
-
-  size_t count (const std::tuple_element<0, for_all_args>::type& value) const {
+  size_t count(const std::tuple_element<0, for_all_args>::type& value) const {
     const derived_type* derived_this = static_cast<const derived_type*>(this);
     derived_this->comm().barrier();
     return ygm::sum(derived_this->local_count(value), derived_this->comm());
